refactor(texts): import interfaces
